+++ src/snapshot.ts
@@ -18,6 +18,7 @@ interface CliOptions {
 	printNotModified: boolean
 	expandDiff: boolean
 	grammar: string[]
+	outdir: string
 }
 
 program
@@ -32,10 +33,11 @@ program
 	.option('--expandDiff', 'produce each diff on two lines prefixed with "++" and "--"', false)
 	.option(
 		'-g, --grammar <grammar>',
-		'Path to a grammar file. Multiple options supported.',
+		'A glob pattern to grammar file(s). Multiple options supported.',
 		array_opt,
 		[],
 	)
+	.option('-o, --outdir <outdir>', 'Specify output directory of testcases', '')
 	.argument(
 		'<testcases...>',
 		'A glob pattern(s) which specifies testcases to run, e.g. "./tests/**/test*.dhall". Quotes are important!',
@@ -58,36 +60,35 @@ async function main(): Promise<ExitCode> {
 
 	const results: ExitCode[] = []
 
-	for (const filename of testCases) {
-		const src = fs.readFileSync(filename, 'utf-8')
-		const scope = filenameToScope(path.basename(filename))
+	for (const test_file of testCases) {
+		const src = fs.readFileSync(test_file, 'utf-8')
+		const scope = filenameToScope(path.basename(test_file))
 		if (scope === undefined) {
-			console.log(chalk.red('ERROR') + " can't run testcase: " + filename)
+			console.log(chalk.red('ERROR') + " can't run testcase: " + test_file)
 			console.log('No scope is associated with the file.')
 			results.push(ExitCode.Failure)
 			continue
 		}
 
 		const tokens = await getVSCodeTokens(registry, scope, src)
-		if (fs.existsSync(filename + '.snap')) {
+		const out_name =
+			options.outdir.length > 0 ? path.join(options.outdir, path.basename(test_file)) : test_file
+		const out_file = out_name + '.snap'
+		if (fs.existsSync(out_file)) {
 			if (options.updateSnapshot) {
-				console.log(
-					chalk.yellowBright('Updating snapshot for ') + chalk.whiteBright(filename + '.snap'),
-				)
+				console.log(chalk.yellowBright('Updating snapshot for ') + chalk.whiteBright(out_file))
 				const text = renderSnapshot(tokens, scope)
-				fs.writeFileSync(filename + '.snap', text, 'utf-8')
+				fs.writeFileSync(out_file, text, 'utf-8')
 				results.push(ExitCode.Success)
 			} else {
-				const snap_text = fs.readFileSync(filename + '.snap', 'utf-8')
+				const snap_text = fs.readFileSync(out_file, 'utf-8')
 				const expectedTokens = unwrap(parseSnap(snap_text))
-				results.push(renderTestResult(filename, expectedTokens, tokens, options))
+				results.push(renderTestResult(out_name, expectedTokens, tokens, options))
 			}
 		} else {
-			console.log(
-				chalk.yellowBright('Generating snapshot ') + chalk.whiteBright(filename + '.snap'),
-			)
+			console.log(chalk.yellowBright('Generating snapshot ') + chalk.whiteBright(out_file))
 			const text = renderSnapshot(tokens, scope)
-			fs.writeFileSync(filename + '.snap', text, 'utf-8')
+			fs.writeFileSync(out_file, text, 'utf-8')
 			results.push(ExitCode.Success)
 		}
 	}